Refactor type definitions and update job handlers to use specific data types for improved type safety
